requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
google-cloud-texttospeech==2.16.3
python-dotenv==1.0.1
